fix validations methods

+++ src/repositorys/displayRepository.ts
@@ -1,8 +1,14 @@
+import mongoose from "mongoose";
 import DisplaySchema from "../schemas/display.ts";
 import Journey from "../schemas/journeys.ts";
+import { ApiError } from "../libs/apiError.ts";
 
 class DisplayRepository {
-    public async fetchOneDisplay(id: any) {
+    public async fetchOneDisplay(id: string) {
+        if (!mongoose.Types.ObjectId.isValid(id)) {
+            throw new ApiError(404, 'Id not valid', 'Não foi possível encontrar esta workspace', { id: id });
+        }
+
         return await DisplaySchema.findById(id);
     }
 
@@ -35,7 +41,11 @@ class DisplayRepository {
         }
     }
 
-    public async updateDisplay(id: any, body: any) {
+    public async updateDisplay(id: string, body: any) {
+        if (!mongoose.Types.ObjectId.isValid(id)) {
+            throw new ApiError(404, 'Id not valid', 'Não foi possível encontrar esta workspace', { id: id });
+        }
+
         return await DisplaySchema.findOneAndUpdate(
             { _id: id },
             {
@@ -45,7 +55,11 @@ class DisplayRepository {
         );
     }
 
-    public async deleteDisplay(id: any) {
+    public async deleteDisplay(id: string) {
+        if (!mongoose.Types.ObjectId.isValid(id)) {
+            throw new ApiError(404, 'Id not valid', 'Não foi possível encontrar esta workspace', { id: id });
+        }
+
         return await DisplaySchema.deleteOne({ _id: id });
     }
 }
