+++ src/repositorys/taskRepository.ts
@@ -1,20 +1,26 @@
+import mongoose from "mongoose";
 import Task from "../schemas/task.ts";
+import { ApiError } from "../libs/apiError.ts";
 
 class TaskRepository {
-    public async fetchOneTask(id: any) {
+    public async fetchOneTask(id: string) {
+        if (!mongoose.Types.ObjectId.isValid(id)) {
+            throw new ApiError(404, 'Id not valid', 'Não foi possível encontrar esta workspace', { id: id });
+        }
+
         return await Task.findById(id);
     }
 
-    public async fetchOneTaskWithDisplay(id: any) {
+    public async fetchOneTaskWithDisplay(id: string) {
+        if (!mongoose.Types.ObjectId.isValid(id)) {
+            throw new ApiError(404, 'Id not valid', 'Não foi possível encontrar esta workspace', { id: id });
+        }
+
         const task = Task.findById(id)
             .populate({ path: "displayId", select: ["-_id", "-__v"] })
             .populate({ path: "tagId", select: ["name", "txt_color_hex"] })
             .exec();
-
-        task.then((resp) => {
-            console.log(resp);
-        });
-
+            
         return await task;
     }
 
@@ -27,7 +33,11 @@ class TaskRepository {
         return await task.save();
     }
 
-    public async updateTask(id: any, body: any) {
+    public async updateTask(id: string, body: any) {
+        if (!mongoose.Types.ObjectId.isValid(id)) {
+            throw new ApiError(404, 'Id not valid', 'Não foi possível encontrar esta workspace', { id: id });
+        }
+
         return await Task.findOneAndUpdate(
             { _id: id },
             { $set: body },
@@ -35,7 +45,11 @@ class TaskRepository {
         );
     }
 
-    public async deleteTask(id: any) {
+    public async deleteTask(id: string) {
+        if (!mongoose.Types.ObjectId.isValid(id)) {
+            throw new ApiError(404, 'Id not valid', 'Não foi possível encontrar esta workspace', { id: id });
+        }
+
         return await Task.deleteOne({ _id: id });
     }
 }
